requirements.txt
# Pretty straight forward django-cms
# Expand more later NBs
django-cms
djangocms-admin-style>=1.5,<1.6
django-treebeard>=4.0,<5.0
djangocms-text-ckeditor>=3.7,<4.0
djangocms-link>=2.5,<2.7
djangocms-icon>=1.4,<1.6
djangocms-style>=2.2,<2.4
djangocms-googlemap>=1.3,<1.5
djangocms-snippet>=2.2,<2.4
djangocms-video>=2.1,<2.4
djangocms-file>=2.3,<2.5
djangocms-picture>=2.3,<2.5
djangocms-bootstrap4>=1.5,<1.7
easy_thumbnails
django-filer>=1.3
# Mainly django and other django packages
Django<2.3
django-classy-tags>=0.9
django-sekizai>=1.0
django-mptt>0.9
django-extensions
django-allauth
django-favicon
djangorestframework
django-shell-plus
django-admin-tools
django-parler
django-taggit
django-taggit-autosuggest
django-meta
django-sortedm2m
djangocms-apphook-setup
django-countries
django-iban
django-money
django-model-utils
djangocms-blog
# Some are just needed
html5lib>=1.0.1
Pillow>=3.0
six
pytz
fstring
lxml
jsonfield
aldryn-apphooks-config
django-widget-tweaks
# Searching the web
google
beautifulsoup4
ordereddict
# Google-Images-Search